Rename upload controller and route
