+++ src/server/upload-and-scan/helpers/upload-stream.js
@@ -1,6 +1,5 @@
-import { PassThrough } from 'node:stream'
 import FileType from 'file-type'
-
+import { PassThrough } from 'node:stream'
 import { Upload } from '@aws-sdk/lib-storage'
 
 import { findS3ContentLength } from '~/src/server/common/helpers/s3/find-s3-content-length'
